requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
boto3==1.28.78
botocore==1.31.78
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.1
Django==4.2.7
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
dnspython==2.4.2
effdet==0.4.1
emoji==2.8.0
exceptiongroup==1.1.3
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
iopath==0.1.10
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.330
langdetect==1.0.9
langsmith==0.0.57
layoutparser==0.3.4
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pinecone-client==2.2.4
portalocker==2.8.2
protobuf==4.25.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
pypdfium2==4.23.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
s3transfer==0.7.0
safetensors==0.4.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
timm==0.9.9
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.10.28
unstructured-inference==0.7.10
unstructured.pytesseract==0.3.12
urllib3==1.26.18
yarl==1.9.2
zipp==3.17.0
